Created User repository.
Added the UserRepository and IUserRepository to the UnitOfWork and IUnitOfWork respectively.

Added additional properties in the user model class.

Altered The Register html to create a custom form.

Created the EmailSender in the utility with a not implemented so that dont get error.

Changed the userRepository's conditional statement for altering the lockoutEnd on the db

Added the the accept attribute to html input for file upload to MenuItem.

Added a style rule to remove the border around the upload field in the site css file.

Added extension validation to the menuItem page and created a error page for the menu item.

Added authorize attribute to all controllers except category(its the home).

Added authorized role checking to all indexs html so that when user enters url directly, access will be denied.

Added authorized role checking to the shared _layout html so that drop downlist will only be rendered if user is authorised.